requirements.txt
discord.py
discord-py-slash-command
python-dotenv
requests
python-dateutil
